Moved location of types.tsx
