+++ todoodlefrontend/src/pages/TodoPage/TodoPage.tsx
@@ -7,8 +7,8 @@ import NavBar from "../../components/NavBar/NavBar"
 import TodoForm from "../../components/Todo/TodoForm"
 import TaskModal from "../../components/Todo/TaskModal"
 import TodoStats from "../../components/Todo/TodoStats"
-import type { Task, Status, Folder } from "../../components/Todo/types"
-import { Statuses, StatusLabels } from "../../components/Todo/types"
+import type { Task, Status, Folder } from "../../types/types"
+import { Statuses, StatusLabels } from "../../types/types"
 import { notesApi, foldersApi } from "../../services/api"
 import { noteToTask, taskToNote, addDefaultColors } from "../../utils/dataTransformers"
 import FolderItem from "../../components/Folder/FolderItem"
